% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ex_Dat}
\alias{Ex_Dat}
\title{Simple example file to illustrate where data should reside, and how to load it simply.}
\format{A .rda data file.}
\usage{
Ex_Dat
}
\description{
The dataset is made up and has no meaningful interpretation
}
\keyword{datasets}
