% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\title{create_template}
\usage{
create_template(directory = "C:/Onetimeonly", template_name, bib.location,
  ShowPDFatLaunch = FALSE)
}
\arguments{
\item{directory}{A character string with the target folder directory This folder will be created if it does not already exist.}

\item{template_name}{The folder, as well as the rmarkdown template, will be given this name.}

\item{bib.location}{This can be a user specified location for a bib file. If omitted, the default bib file will be included. Note this command will create a renamed copy in the provided directory root.}

\item{ShowPDFatLaunch}{FALSE by default, this will launch the default PDF.}
}
\value{
If launch_template is TRUE, the template Rmarkdown file will be opened in a new R session. This can then be used as a starting point for an academic research paper.
}
\description{
This function builds a template directory  with all the necessary files, as well as a rmarkdown script which is ready to edit.
The result is an Elsevier ready document that conforms to most academic journal standards.
In the directory folder provided, the user will have the tex files, as well as all relevant figure folders created in R. These should not be renamed or moved.
Prior to submitting a paper, the user should zip the template_name.tex file, together with the figure files and Tex folder, and send it through to journal editors.
Feel free to edit any of the settings in the template_name.tex file to conform to particular preferences (e.g. setting margins, etc.)
}
\examples{
\donttest{create_template(directory = "C:/Temp", template_name = "Project")}
}
